Related Search quick fix * Update Inquisition and fix the related search feature when there is no search index.
